refactor: restructure project into monorepo with keytool and server components
