+++ .github/workflows/basic-branch-build.yml
@@ -16,5 +16,10 @@ jobs:
       with:
         go-version: '1.21'
 
-    - name: Build
+    - name: Build Keytool
+      working-directory: keytool
+      run: go build -v ./...
+
+    - name: Build Server
+      working-directory: server
       run: go build -v ./...
