+++ .gitignore
@@ -24,5 +24,6 @@ go.work.sum
 # env file
 .env
 
-#
+# apple and diverse stuff
 .idea
+.DS_Store
