--- go.mod
@@ -1,5 +1,0 @@
-module oauth2-task
-
-go 1.24.2
-
-require github.com/golang-jwt/jwt/v5 v5.2.2
