--- pkg/keys/go.mod
@@ -1,3 +1,0 @@
-module github.com/wunderkind2k1/auth-server-task/pkg/keys
-
-go 1.21
