--- pkg/keys/keys/cddcbf9fe23b31ad.public.pem
@@ -1,8 +1,0 @@
------BEGIN RSA PUBLIC KEY-----
-MIIBCgKCAQEAzdy/n+I7Ma2QZTXqvtPbP2PmZmcVOEfxzW3ZMrNLlDBMMiUWF025
-LpNeL1SpeHBVNPMQuHqggGRPxsbep5V3DLor6ZojS/de1ImPMmUWO6PSalZtCMX1
-lpL6mGxxMGFqteXpAf7IXFV/M+yEMwbxuIIcPd8bUdfg8VVISxjGscaRutXOdMyn
-ze9rzt+1zu9qNsgV5Q/K19K4y+z1os+pkjHCCW3zicdoHYMbLy4Pw0A7ys6s7ZNp
-12dbr45yN9rlheefkgjdZy1Z5xnxA+xXvdJaQ/VPxbSI5OroUQtg8n5h+L2ZnnUv
-54j0BHJI2/Z8Rpcrl/d/LtMU8nr7MiVMRwIDAQAB
------END RSA PUBLIC KEY-----
