+++ server/internal/token/generator.go
@@ -1,0 +1,48 @@
+// Package token provides JWT token generation and key management functionality.
+// It implements token generation using RSA private keys and handles key loading
+// from files. The package follows JWT standards for token creation and signing.
+package token
+
+import (
+	"crypto/rsa"
+	"log/slog"
+	"time"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+// Generator handles JWT token generation
+type Generator struct {
+	privateKey *rsa.PrivateKey
+}
+
+// NewGenerator creates a new token generator
+func NewGenerator(privateKey *rsa.PrivateKey) *Generator {
+	return &Generator{
+		privateKey: privateKey,
+	}
+}
+
+// GenerateToken creates a new JWT token for the given username
+func (g *Generator) GenerateToken(username string) (string, error) {
+	// Create the Claims
+	claims := jwt.RegisteredClaims{
+		ExpiresAt: jwt.NewNumericDate(time.Now().Add(1 * time.Hour)),
+		IssuedAt:  jwt.NewNumericDate(time.Now()),
+		NotBefore: jwt.NewNumericDate(time.Now()),
+		Issuer:    "auth-server",
+		Subject:   username,
+	}
+
+	// Create token with claims
+	token := jwt.NewWithClaims(jwt.SigningMethodRS256, claims)
+
+	// Generate encoded token
+	tokenString, err := token.SignedString(g.privateKey)
+	if err != nil {
+		slog.Error("failed to sign JWT token", "error", err)
+		return "", err
+	}
+
+	return tokenString, nil
+}
